Fix erreur si le panier était vide

+++ js/cart.js
@@ -18,40 +18,42 @@
         // Affiche tout les produits présent dans le panier
         let products = [];
         (showCart = () => {
-            productLocalStorage.forEach((print, i) => {
-                let id = productLocalStorage[i]._id;
-                let productObject = findObject(id);
-                let price = productObject.price * parseInt(productLocalStorage[i].qty);
-                
-                // Stockage des id produits présent dans le panier
-                let productsId = [productLocalStorage[i]._id];
-                products.push(productsId);
-
-                // Insertion dans la page 
-                let cart = document.getElementById("cart__items");
-                cart.innerHTML += `
-                        <article class="cart__item" data-id="${productLocalStorage[i]._id}" data-color="${productLocalStorage[i].colors}">
-                            <div class="cart__item__img">
-                                <img src="${productObject.imageUrl}" alt="${productObject.altTxt}">
-                            </div>
-                            <div class="cart__item__content">
-                                <div class="cart__item__content__description">
-                                    <h2>${productObject.name}</h2>
-                                    <p>${productLocalStorage[i].colors}</p>
-                                    <p class="priceProduct">${price}€</p>
+            if (localStorage.getItem('product')){
+                productLocalStorage.forEach((print, i) => {
+                    let id = productLocalStorage[i]._id;
+                    let productObject = findObject(id);
+                    let price = productObject.price * parseInt(productLocalStorage[i].qty);
+                    
+                    // Stockage des id produits présent dans le panier
+                    let productsId = [productLocalStorage[i]._id];
+                    products.push(productsId);
+    
+                    // Insertion dans la page 
+                    let cart = document.getElementById("cart__items");
+                    cart.innerHTML += `
+                            <article class="cart__item" data-id="${productLocalStorage[i]._id}" data-color="${productLocalStorage[i].colors}">
+                                <div class="cart__item__img">
+                                    <img src="${productObject.imageUrl}" alt="${productObject.altTxt}">
                                 </div>
-                                <div class="cart__item__content__settings">
-                                    <div class="cart__item__content__settings__quantity">
-                                        <p>Qté : </p>
-                                        <input type="number" class="itemQuantity" name="itemQuantity" min="1" max="100" value="${productLocalStorage[i].qty}">
+                                <div class="cart__item__content">
+                                    <div class="cart__item__content__description">
+                                        <h2>${productObject.name}</h2>
+                                        <p>${productLocalStorage[i].colors}</p>
+                                        <p class="priceProduct">${price}€</p>
                                     </div>
-                                    <div class="cart__item__content__settings__delete">
-                                    <p class="deleteItem">Supprimer</p>
+                                    <div class="cart__item__content__settings">
+                                        <div class="cart__item__content__settings__quantity">
+                                            <p>Qté : </p>
+                                            <input type="number" class="itemQuantity" name="itemQuantity" min="1" max="100" value="${productLocalStorage[i].qty}">
+                                        </div>
+                                        <div class="cart__item__content__settings__delete">
+                                        <p class="deleteItem">Supprimer</p>
+                                        </div>
                                     </div>
                                 </div>
-                            </div>
-                        </article>`
-            })
+                            </article>`
+                })
+            }
         })();
 
         // Afficher le coût total et le nombre d'objet du panier
